schema updates.

+++ api/sql/schema/00008_ownerships.sql
@@ -4,9 +4,9 @@ SELECT 'up SQL query';
 CREATE TABLE ownership
 (
     id                    SERIAL PRIMARY KEY,
-    unit_id               INTEGER         NOT NULL REFERENCES units (id),
-    owner_id              INTEGER         NOT NULL REFERENCES owners (id),
-    part                  DECIMAL(20, 18) NOT NULL,
+    unit_id               INTEGER NOT NULL REFERENCES units (id),
+    owner_id              INTEGER NOT NULL REFERENCES owners (id),
+    part                  NUMERIC NOT NULL,
     start_date            TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     end_date              TIMESTAMP NULL, -- NULL means current ownership
     is_active             BOOLEAN   DEFAULT TRUE,
